Parallelized radial distribution function (#1)
* :sparkles: :recycle:

Lots of changes, first, trying to refactor de RDF to make it parallel.
Then, factoring out the structure factor subroutines to make it more
modular.

* :recycle:

Refactoring done, encapsulated most of the observables into their
respective subroutines. Added a global parameter to control whether the
structure factor needs to be computed or not.

* :sparkles:

It seems to be working as expected, the RDF seems to be appropriately
normalized.

* :recycle:

Use Newton's third law and avoid doing double computations.

* :art:

Add some ways of knowing that the program has finished.

* :art:

It is better to leave the thermalization steps parameter as a runtime
parameter, instead of hard-coding it.

* :art:

Use less averaging cycles for the final observables.